Fixed issue where CORS connections could not be made in Internet Explorer
If the http-bind URL was on a different domain to the page running the
plugin, $.ajax would fail in Internet Explorer and a connection to the
server would not be made.

I've added some code to intercept $.ajax when coming from IE8 and above
(it doesn't work in anything lower than IE8 unfortunately) and then send
it via IE's built in XDomainRequest function.

The code originated from the jQuery forum at
http://forum.jquery.com/topic/cross-domain-ajax-and-ie#14737000002203097,
although I slightly modified it to better fit this purpose.